[gitignore] Exclude IntelliJ files (.idea)
IntelliJ's .idea/ folder is not meant to be shared. I've seen too many devs
getting frustrated when the files are shared, specially when devs tend to
adopt different versions of the IDE (betas, previews, RC, ect).

Until JetBrains comes up to their senses and make shareable config files
just as Eclispe and VS Code do it is best to avoid sharing the files.