The plugin applies the detekt plugin and provides sensible defaults
This project configures the detekt plugin (io.gitlab.arturbosch.detekt:1.17.1).

By loading this pluging into a gradle project:

  plugins {
      id 'a8c-apply-style-gradle-plugin' version '0.0.1'
  }

The plugin sets the project as if the following was done:

  allprojects {
    apply plugin: 'io.gitlab.arturbosch.detekt'

    detekt {
        // If no detekt file is provided by the project a default one
        // is provided by the plugin
        config = files("${project.rootDir}/config/detekt/detekt.yml")

        autoCorrect = false
        buildUponDefaultConfig = false
        disableDefaultRuleSets = false
        ignoreFailures = false
        parallel = false
        debug = false
        reports {
            html.enabled = true
            xml.enabled = true
            txt.enabled = true
        }

        // If the project has baseline file use it
        def baselineFile = file("${project.rootDir}/config/detekt/baseline.xml")
        if (baselineFile.exists()) {
           baseline = baselineFile
        }
    }
  }

It is possible to test the plugin from another project with:

  A8C_DEBUG=1 ./gradlew --include-build ../apply-style-gradle-plugin --quiet WordPress:detekt

The environment variable `A8C_DEBUG` is used for enabling debug only for
the plugin.

The plugin adds the task `extractDetektConf` that can be used for
extracting the configuration bundled in the plugin. Ideally we don't need
to extract the configuration at all.

By default the extraction is based on the gradle config defined under
`detekt.config` if there's no such config then file will be extracted
under `${project.buildDir/detekt.yml}`.